+++ changelog.md
@@ -9,6 +9,7 @@
 * **Fix** - Added better customizer default handling via the `layers_customizer_control_defaults` hook
 * **Fix** - Payment method block alignment no longer has a margin on the left
 * **Fix** - Fixed pagination location on the post widget
+* **Fix** - Gutter option on all widgets with masonry active now works
 * **Tweak** - Header cart background color has changed for a hash value to a transparent rgba background color for better handling of different header colors
 * **Tweak** - Improved spacing of the comment form block as well as a font-size decrease for "Leave a Reply"
 * **Tweak** - Gave copyright border-color rgba (same reason as header cart)